template gallery refactoring
